requirements.txt
gradio
llama-cpp-python
